// src/app/models/Footer.ts
export interface Footer{
    id: number;
    name_shop: string;
    logo: string;
    description: string;
    name_ecommerce_1: string;
    url_ecommerce_1: string;
    name_ecommerce_2: string;
    url_ecommerce_2: string;
    name_ecommerce_3: string;
    url_ecommerce_3: string;
    phone_1: string;
    phone_2: string;
    email: string;
    address: string;
    created_at: string;
    updated_at: string;
    status: boolean;
}